wip: ingest interface
