+++ pychunkedgraph/backend/chunkedgraph_edits.py
@@ -11,7 +11,7 @@ from multiwrapper import multiprocessing_utils as mu
 from . import flatgraph_utils
 from .chunkedgraph_utils import combine_cross_chunk_edge_dicts
 from .utils import column_keys, serializers
-from .utils.helpers import get_bounding_box
+from .utils.general import get_bounding_box
 from .utils.edge_utils import filter_fake_edges
 from .utils.edge_utils import map_edges_to_chunks
 from .utils.edge_utils import get_linking_edges
